+++ examples/Spread/Spread.ino
@@ -1,5 +1,5 @@
 // M16 Spread example
-// Spread reads the oscillator 3 times, each at a different frequency and sums the output
+// The Spread function reads the oscillator 3 times, each at a different frequency and sums the output
 // Useful for oscillator detune or chordal effects
 // ESP8266 struggles with spread, try using window transform or osc morph instead
 #include "M16.h" 
@@ -49,7 +49,7 @@ void loop() {
 * Always finish with i2s_write_samples()
 */
 void audioUpdate() {
-  int16_t leftVal = (filter.nextLPF(aOsc1.next()) * vol)>>10;
-  int16_t rightVal = leftVal;
+  int32_t leftVal = (filter.nextLPF(aOsc1.next()) * vol)>>10;
+  int32_t rightVal = leftVal;
   i2s_write_samples(leftVal, rightVal);
 }
