+++ examples/Waveshaper/Waveshaper.ino
@@ -61,7 +61,7 @@ void loop() {
 * Always finish with i2s_write_samples()
 */
 void audioUpdate() {
-  int16_t leftVal = (effect1.waveShaper(aOsc1.next(), shapeMixVal) * vol)>>10;
-  int16_t rightVal = leftVal;
+  int32_t leftVal = (effect1.waveShaper(aOsc1.next(), shapeMixVal) * vol)>>10;
+  int32_t rightVal = leftVal;
   i2s_write_samples(leftVal, rightVal);
 }
